+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix Language cleaning function that wasn't working.
 - update Report by downgrading PublishedDate's issue to a warning.
 - add a new PublishedDate format ("2006-01-02T15:04:05")
+- add a new Series format ("[<Series>-<SeriesIndex>] <SeriesTitle>")
 
 ## [0.4.1] - 2023-05-19
 - fix lacking book.Path information after edition.